clean-up (it should compile)
