+++ src/accessors-geog.cpp
@@ -1,5 +1,7 @@
 #include <s2geography.h>
+#include <s2geography/geography.h>
 
+#include "creation.hpp"
 #include "geography.hpp"
 #include "pybind11.hpp"
 
@@ -12,15 +14,12 @@ const double EARTH_RADIUS_METERS = 6371.01 * 1000;
 PyObjectGeography centroid(PyObjectGeography a) {
     const auto& a_ptr = a.as_geog_ptr()->geog();
     auto s2_point = s2geog::s2_centroid(a_ptr);
-    std::unique_ptr<Point> point =
-        make_geography<s2geog::PointGeography, spherely::Point>(s2_point);
-    return PyObjectGeography::from_geog(std::move(point));
+    return make_py_geography<s2geog::PointGeography>(s2_point);
 }
 
 PyObjectGeography boundary(PyObjectGeography a) {
     const auto& a_ptr = a.as_geog_ptr()->geog();
     auto s2_obj = s2geog::s2_boundary(a_ptr);
-    // TODO return specific subclass
     auto geog_ptr = std::make_unique<spherely::Geography>(std::move(s2_obj));
     return PyObjectGeography::from_geog(std::move(geog_ptr));
 }
@@ -28,7 +27,7 @@ PyObjectGeography boundary(PyObjectGeography a) {
 PyObjectGeography convex_hull(PyObjectGeography a) {
     const auto& a_ptr = a.as_geog_ptr()->geog();
     auto s2_obj = s2geog::s2_convex_hull(a_ptr);
-    auto geog_ptr = std::make_unique<spherely::Polygon>(std::move(s2_obj));
+    auto geog_ptr = std::make_unique<spherely::Geography>(std::move(s2_obj));
     return PyObjectGeography::from_geog(std::move(geog_ptr));
 }
