+++ src/creation.cpp
@@ -1,3 +1,5 @@
+#include "creation.hpp"
+
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <s2/s2latlng.h>
@@ -24,25 +26,27 @@ using namespace spherely;
 // ---- S2geometry object creation functions.
 //
 
-using PointVec = std::vector<Point *>;
+// using PointVec = std::vector<Point *>;
 using LatLngVec = std::vector<std::pair<double, double>>;
 
-// Used in Geography constructors to get a point either from a tuple of
-// coordinates or an existing Point object.
-S2Point to_s2point2(const std::pair<double, double> &vertex) {
+S2Point make_s2point(double lng, double lat) {
+    return S2LatLng::FromDegrees(lat, lng).ToPoint();
+}
+
+S2Point make_s2point(const std::pair<double, double> &vertex) {
     return S2LatLng::FromDegrees(vertex.first, vertex.second).ToPoint();
 }
 
-S2Point to_s2point2(const Point *vertex) {
-    return vertex->s2point();
-}
+// S2Point to_s2point2(const Point *vertex) {
+//     return vertex->s2point();
+// }
 
 template <class V>
-std::vector<S2Point> to_s2points2(const std::vector<V> &vertices) {
+std::vector<S2Point> make_s2points(const std::vector<V> &vertices) {
     std::vector<S2Point> points(vertices.size());
 
     auto func = [](const V &vertex) {
-        return to_s2point(vertex);
+        return make_s2point(vertex);
     };
 
     std::transform(vertices.begin(), vertices.end(), points.begin(), func);
@@ -87,136 +91,172 @@ std::unique_ptr<S2Loop> make_s2loop(const std::vector<V> &vertices, bool check =
 }
 
 //
-// ---- S2geometry / S2Geography / Spherely object wrapper utility functions
+// Helpers for cloning s2geography objects.
 //
 
-/*
-** Helper to wrap a S2Geometry object as a Spherely C++ Geography object.
-**
-** There are two nested wrap levels:
-** spherely::Geography > s2geography::Geography > s2geometry object
-**
-** @tparam T The Spherely Geography type
-** @tparam S The S2Geometry type
-*/
-template <class T, class S, std::enable_if_t<std::is_base_of_v<s2geog::Geography, T>, bool> = true>
-inline std::unique_ptr<Geography> make_geography2(S &&s2_obj) {
-    using S2GeographyType = T;
-
-    auto s2geog_ptr = std::make_unique<S2GeographyType>(std::forward<S>(s2_obj));
-    return std::make_unique<Geography>(std::move(s2geog_ptr));
+template <class T, std::enable_if_t<std::is_base_of_v<s2geog::Geography, T>, bool> = true>
+const T *downcast(const s2geog::Geography &geog) {
+    return dynamic_cast<const T *>(&geog);
 }
 
-/*
-** Helper to create Spherely Python Geography objects directly from S2Geometry objects.
-**
-** @tparam T The Spherely Geography type
-** @tparam S The S2Geometry type
-*/
-template <class T, class S, std::enable_if_t<std::is_base_of_v<s2geog::Geography, T>, bool> = true>
-inline std::unique_ptr<T> make_py_geography(S &&s2_obj) {
-    auto geog = make_geography2<T>(std::forward<S>(s2_obj));
-    return PyObjectGeography::from_geog(std::move(geog));
+// Use some dirty worarounds and would probably better to be implemented
+// via a `s2geography::Geography::clone` virtual method.
+template <class T, std::enable_if_t<std::is_same_v<T, s2geog::PointGeography>, bool> = true>
+std::unique_ptr<s2geog::Geography> _clone_s2geography(const s2geog::Geography &geog) {
+    const auto &points = static_cast<const T &>(geog).Points();
+    return std::make_unique<T>(points);
 }
+
+template <class T, std::enable_if_t<std::is_same_v<T, s2geog::PolylineGeography>, bool> = true>
+std::unique_ptr<s2geog::Geography> _clone_s2geography(const s2geog::Geography &geog) {
+    const auto &polylines = static_cast<const T &>(geog).Polylines();
+    std::vector<std::unique_ptr<S2Polyline>> polylines_copy(polylines.size());
+
+    auto copy_polyline = [](const std::unique_ptr<S2Polyline> &polyline) {
+        return std::unique_ptr<S2Polyline>(polyline->Clone());
+    };
+
+    std::transform(polylines.begin(), polylines.end(), polylines_copy.begin(), copy_polyline);
+
+    return std::make_unique<T>(std::move(polylines_copy));
+}
+
+template <class T, std::enable_if_t<std::is_same_v<T, s2geog::PolygonGeography>, bool> = true>
+std::unique_ptr<s2geog::Geography> _clone_s2geography(const s2geog::Geography &geog) {
+    const auto &poly = static_cast<const T &>(geog).Polygon();
+    std::unique_ptr<S2Polygon> poly_ptr(poly->Clone());
+    return std::make_unique<T>(std::move(poly_ptr));
+}
+
+template <class T, std::enable_if_t<std::is_same_v<T, s2geog::GeographyCollection>, bool> = true>
+std::unique_ptr<s2geog::Geography> _clone_s2geography(const s2geog::Geography &geog) {
+    const auto &features = static_cast<const T &>(geog).Features();
+    std::vector<std::unique_ptr<s2geog::Geography>> features_copy;
+    features_copy.reserve(features.size());
+
+    for (const auto &feature_ptr : features) {
+        features_copy.push_back(clone_s2geography(*feature_ptr));
+    }
+
+    return std::make_unique<T>(std::move(features_copy));
+}
+
+std::unique_ptr<s2geog::Geography> spherely::clone_s2geography(const s2geog::Geography &geog) {
+    if (const auto *ptr = downcast<s2geog::PointGeography>(geog); ptr) {
+        return _clone_s2geography<s2geog::PointGeography>(*ptr);
+    } else if (const auto *ptr = downcast<s2geog::PolylineGeography>(geog); ptr) {
+        return _clone_s2geography<s2geog::PolylineGeography>(*ptr);
+    } else if (const auto *ptr = downcast<s2geog::PolygonGeography>(geog); ptr) {
+        return _clone_s2geography<s2geog::PolygonGeography>(*ptr);
+    } else if (const auto *ptr = downcast<s2geog::GeographyCollection>(geog); ptr) {
+        return _clone_s2geography<s2geog::GeographyCollection>(*ptr);
+    } else {
+        throw std::runtime_error("invalid Geography type");
+    }
+}
+
+//
+// ---- S2geometry / S2Geography / Spherely object wrapper utility functions
+//
 
 //
 // ---- Spherely C++ Geography creation functions
 //
 
-std::unique_ptr<Geography> create_point2(double lng, double lat) {
-    return make_geography2<s2geog::PointGeography>(to_s2point2(std::make_pair(lat, lng)));
-}
+// std::unique_ptr<Geography> create_point2(double lng, double lat) {
+//     return make_geography<s2geog::PointGeography>(to_s2point2(std::make_pair(lat, lng)));
+// }
 
-template <class V>
-std::unique_ptr<Geography> create_multipoint2(const std::vector<V> &pts) {
-    return make_geography2<s2geog::PointGeography>(to_s2points(pts));
-}
+// template <class V>
+// std::unique_ptr<Geography> create_multipoint2(const std::vector<V> &pts) {
+//     return make_geography2<s2geog::PointGeography>(to_s2points(pts));
+// }
 
-template <class V>
-std::unique_ptr<Geography> create_linestring2(const std::vector<V> &pts) {
-    auto s2points = to_s2points(pts);
-    auto polyline_ptr = std::make_unique<S2Polyline>(s2points);
+// template <class V>
+// std::unique_ptr<Geography> create_linestring2(const std::vector<V> &pts) {
+//     auto s2points = to_s2points(pts);
+//     auto polyline_ptr = std::make_unique<S2Polyline>(s2points);
 
-    return make_geography2<s2geog::PolylineGeography>(std::move(polyline_ptr));
-}
+//     return make_geography2<s2geog::PolylineGeography>(std::move(polyline_ptr));
+// }
 
-std::unique_ptr<Geography> create_multilinestring2(const std::vector<LineString *> &lines) {
-    std::vector<std::unique_ptr<S2Polyline>> polylines(lines.size());
+// std::unique_ptr<Geography> create_multilinestring2(const std::vector<LineString *> &lines) {
+//     std::vector<std::unique_ptr<S2Polyline>> polylines(lines.size());
 
-    auto func = [](const LineString *line_ptr) {
-        S2Polyline *cloned_ptr(line_ptr->s2polyline().Clone());
-        return std::make_unique<S2Polyline>(std::move(*cloned_ptr));
-    };
+//     auto func = [](const LineString *line_ptr) {
+//         S2Polyline *cloned_ptr(line_ptr->s2polyline().Clone());
+//         return std::make_unique<S2Polyline>(std::move(*cloned_ptr));
+//     };
 
-    std::transform(lines.begin(), lines.end(), polylines.begin(), func);
+//     std::transform(lines.begin(), lines.end(), polylines.begin(), func);
 
-    return make_geography2<s2geog::PolylineGeography>(std::move(polylines));
-}
+//     return make_geography2<s2geog::PolylineGeography>(std::move(polylines));
+// }
 
-template <class V>
-std::unique_ptr<Geography> create_multilinestring2(const std::vector<std::vector<V>> &lines) {
-    std::vector<std::unique_ptr<S2Polyline>> polylines(lines.size());
+// template <class V>
+// std::unique_ptr<Geography> create_multilinestring2(const std::vector<std::vector<V>> &lines) {
+//     std::vector<std::unique_ptr<S2Polyline>> polylines(lines.size());
 
-    auto func = [](const std::vector<V> &pts) {
-        auto s2points = to_s2points(pts);
-        return std::make_unique<S2Polyline>(s2points);
-    };
+//     auto func = [](const std::vector<V> &pts) {
+//         auto s2points = to_s2points(pts);
+//         return std::make_unique<S2Polyline>(s2points);
+//     };
 
-    std::transform(lines.begin(), lines.end(), polylines.begin(), func);
+//     std::transform(lines.begin(), lines.end(), polylines.begin(), func);
 
-    return make_geography2<s2geog::PolylineGeography>(std::move(polylines));
-}
+//     return make_geography2<s2geog::PolylineGeography>(std::move(polylines));
+// }
 
-template <class V>
-std::unique_ptr<Geography> create_polygon2(
-    const std::vector<V> &shell,
-    const std::optional<std::vector<std::vector<V>>> &holes) {
-    std::vector<std::unique_ptr<S2Loop>> loops;
-    loops.push_back(make_s2loop(shell, false));
+// template <class V>
+// std::unique_ptr<Geography> create_polygon2(
+//     const std::vector<V> &shell,
+//     const std::optional<std::vector<std::vector<V>>> &holes) {
+//     std::vector<std::unique_ptr<S2Loop>> loops;
+//     loops.push_back(make_s2loop(shell, false));
 
-    if (holes.has_value()) {
-        for (const auto &ring : holes.value()) {
-            loops.push_back(make_s2loop(ring, false));
-        }
-    }
+//     if (holes.has_value()) {
+//         for (const auto &ring : holes.value()) {
+//             loops.push_back(make_s2loop(ring, false));
+//         }
+//     }
 
-    auto polygon_ptr = std::make_unique<S2Polygon>();
-    polygon_ptr->set_s2debug_override(S2Debug::DISABLE);
-    polygon_ptr->InitNested(std::move(loops));
+//     auto polygon_ptr = std::make_unique<S2Polygon>();
+//     polygon_ptr->set_s2debug_override(S2Debug::DISABLE);
+//     polygon_ptr->InitNested(std::move(loops));
 
-    // Note: this also checks each loop of the polygon
-    if (!polygon_ptr->IsValid()) {
-        std::stringstream err;
-        S2Error s2err;
-        err << "polygon is not valid: ";
-        polygon_ptr->FindValidationError(&s2err);
-        err << s2err.text();
-        throw py::value_error(err.str());
-    }
+//     // Note: this also checks each loop of the polygon
+//     if (!polygon_ptr->IsValid()) {
+//         std::stringstream err;
+//         S2Error s2err;
+//         err << "polygon is not valid: ";
+//         polygon_ptr->FindValidationError(&s2err);
+//         err << s2err.text();
+//         throw py::value_error(err.str());
+//     }
 
-    return make_geography2<s2geog::PolygonGeography>(std::move(polygon_ptr));
-}
+//     return make_geography2<s2geog::PolygonGeography>(std::move(polygon_ptr));
+// }
 
-std::unique_ptr<GeographyCollection> create_collection2(const std::vector<Geography *> &features) {
-    std::vector<std::unique_ptr<s2geog::Geography>> features_copy;
-    features_copy.reserve(features.size());
+// std::unique_ptr<GeographyCollection> create_collection2(const std::vector<Geography *> &features)
+// {
+//     std::vector<std::unique_ptr<s2geog::Geography>> features_copy;
+//     features_copy.reserve(features.size());
 
-    for (const auto &feature_ptr : features) {
-        features_copy.push_back(clone_s2geography(feature_ptr->geog()));
-    }
+//     for (const auto &feature_ptr : features) {
+//         features_copy.push_back(clone_s2geography(feature_ptr->geog()));
+//     }
 
-    return std::make_unique<GeographyCollection>(
-        std::make_unique<s2geog::GeographyCollection>(std::move(features_copy)));
-    ;
-}
+//     return std::make_unique<GeographyCollection>(
+//         std::make_unique<s2geog::GeographyCollection>(std::move(features_copy)));
+//     ;
+// }
 
 //
 // ---- Spherely Python Geography creation functions
 //
 
 PyObjectGeography point(double longitude, double latitude) {
-    auto geog = create_point2(longitude, latitude);
-    return PyObjectGeography::from_geog(std::move(geog));
+    return make_py_geography<s2geog::PointGeography>(make_s2point(longitude, latitude));
 }
 
 //
